+++ js/packages/phoenix-otel/src/createNoOpProvider.ts
@@ -1,0 +1,10 @@
+import { NodeTracerProvider } from "@opentelemetry/sdk-trace-node";
+
+/**
+ * For dry runs we create a provider that doesn't export traces.
+ */
+export function createNoOpProvider() {
+  const provider = new NodeTracerProvider({});
+
+  return provider;
+}