+++ js/packages/phoenix-otel/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @arizeai/phoenix-otel
 
+## 0.2.0
+
+### Minor Changes
+
+- de6f111: refactor to use phoenix-otel across the client
+
+## 0.1.0
+
+### Minor Changes
+
+- 08c7419: add support for passing instrumentation to register"
+
 ## 0.0.1
 
 ### Patch Changes